test(use-online): migrate interaction tests to browser mode
